Adding function to pad start of number string
JS pad start is not supported on the brightsign.
Ran into issues where I was using i as the counter in the padNumber function but i was also the counter in the loop through the classes, and when the function was called, it was passing also the counter variable somehow? Not sure why it does that.